+++ kernel_sum.go
@@ -17,12 +17,12 @@ func (k *sumKernel) Offset() int {
 	return (k.size - 1) / 2
 }
 
-func (k *sumKernel) Apply(in *image.Gray, x, y int) int {
+func (k *sumKernel) Apply(in *image.Gray, x, y int) float64 {
 	sum := 0
 	for i := -k.Offset(); i <= k.Offset(); i++ {
 		for j := -k.Offset(); j <= k.Offset(); j++ {
 			sum += int(in.GrayAt(x+i, y+j).Y)
 		}
 	}
-	return sum
+	return float64(sum)
 }
